Update Conditional navbars and create config for navbar
